chore: add utility functions for HTTP2 testing
